issue #114 refactor: Preference 저장소를 :core:datastore 로 이동, AuthModule 분할
PreferenceRepository/Impl/Keys 를 :core:datastore 로 이동하고, DataStore/
PreferenceRepository 제공을 DataStoreModule 로 신설한다. AuthModule 은
AuthInterceptor/AuthAuthenticator(network 파트)만 남기고, datastore 파트를
분리했다. 같은 패키지라 import 없이 참조하던 Auth* 에는 core.datastore
import 를 명시한다.
