+++ app/build.gradle.kts
@@ -49,6 +49,7 @@ dependencies {
     implementation(project(":core:model"))
     implementation(project(":core:common"))
     implementation(project(":core:ui"))
+    implementation(project(":core:datastore"))
 
     implementation(libs.core.ktx)
     implementation(libs.activity.compose)
