+++ core/datastore/src/main/java/cord/eoeo/momentwo/core/datastore/PreferenceRepository.kt
@@ -1,4 +1,4 @@
-package cord.eoeo.momentwo.data.authentication
+package cord.eoeo.momentwo.core.datastore
 
 interface PreferenceRepository {
     suspend fun storeAccessToken(accessToken: String)
